Fix run_dir handling
This fix allows the use of alternative TensorFlow SummaryWriter's (e.g., GCP, S3, etc). Using pathlib.Path, it changes incoming --run-dir=s3://path to s3:/path (Note the single slash) and the SummaryWriter check fails to match due to the single slash.

The fix is to not wrap in a pathlib.Path and do inline concatenation using os.path.join