Add disable_amp flag to OCP calculator
Having AMP on often leads to trouble during inference (rounding errors, energies not changing during relaxations, etc). This PR adds a flag to the calculator to force it off, and also defaults that to True. This is done using the trainer.scaling=None trick, but we could also overwrite the config if that was cleaner

Users who know what they're doing can use disable_amp=False, but the default for new users who aren't aware of issues should be AMP off.