Adding omitted dependencies to `fairchem-core`'s `pyproject.toml`
# Purpose

As noted in the [installation guide](https://fair-chem.github.io/core/install.html):

> Install torch_geometric and the torch_scatter, torch_sparse, and torch_cluster optional dependencies similarly by selecting the appropriate versions in the official PyG docs

Based on these instructions, it appears that `torch_geometric` is a requirement for `fairchem-core` yet is not included in the list of required dependencies in `pyproject.toml`. Additionally, there is no optional set of dependencies in `pyproject.toml` for `torch_scatter`, `torch_sparse`, or `torch_cluster`.

While the installation guide makes it very clear that these packages are to be installed by going to the PyG website (and this is indeed the recommended approach), it would still be beneficial to have these packages specified in `pyproject.toml` for two reasons:
1. If there ever ends up being a dependency issue where a new release of `torch_geometric` breaks `fairchem-core`, you can put a downward version pin `<=` in `pyproject.toml`, which will ensure that the user gets the compatible version when they go to install `torch_geometric` from the PyG website's instructions.
2. Packages using `fairchem-core` as a dependency will not have to each of the aforementioned dependencies to their package when they are, in reality, first-class dependencies of `fairchem-core`. This would be very helpful for maintainers (e.g. me!) because they otherwise have to be regularly updating/testing four additional dependencies they may not otherwise need to keep track of. For additional context, see my post [here](https://github.com/Quantum-Accelerators/quacc/pull/2680#discussion_r1976673049).

# Change

I have put `torch_geometric` in the required dependencies and made a new optional dependency block called `pyg` for the others.